Backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
chromadb
sentence-transformers
numpy
python-dotenv